requirements.txt
environ
amqp==2.5.0
aniso8601==6.0.0
args==0.1.0
asn1crypto==0.24.0
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
clint==0.5.1
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-allauth==0.39.1
django-ckeditor==5.7.1
django-environ==0.4.5
django-import-export==1.2.0
django-js-asset==1.2.2
django-tables2==2.0.6
django-widget-tweaks==1.4.3
dna==0.0.7
docker-py==1.10.6
docker-pycreds==0.4.0
docopt==0.6.2
et-xmlfile==1.0.1
Flask==1.0.3
flask-mongoengine==0.9.5
Flask-RESTful==0.3.7
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
kombu==4.6.0
Logbook==1.4.3
MarkupSafe==1.1.1
mongoengine==0.17.0
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
pycparser==2.19
pymongo==3.8.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1
raven==6.10.0
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
sh==1.12.14
six==1.12.0
sqlparse==0.3.0
structlog==19.1.0
tablib==0.13.0
tmdbsimple==2.2.0
tmdbv3api==1.3.2
urllib3==1.25.2
vine==1.3.0
websocket-client==0.56.0
Werkzeug==0.15.4
wget==3.2
WTForms==2.2.1
xlrd==1.2.0
xlwt==1.3.0
